:recycle: refactor: Extract AppFrame from App.
