+++ imports/ui/App.tsx
@@ -2,10 +2,7 @@ import React from 'react';
 
 import {BrowserRouter} from 'react-router-dom';
 
-import {useTracker} from 'meteor/react-meteor-data';
-
 import {
-	makeStyles,
 	MuiThemeProvider,
 	createMuiTheme,
 	responsiveFontSizes,
@@ -13,66 +10,21 @@ import {
 
 import {SnackbarProvider} from 'notistack';
 
-import LinearProgress from '@material-ui/core/LinearProgress';
-
 import CssBaseline from '@material-ui/core/CssBaseline';
 
 import {LocalizationProvider} from '@material-ui/pickers';
 import AdapterDateFns from '@material-ui/pickers/adapter/date-fns';
 
-import {all} from '@iterable-iterator/reduce';
-import {map} from '@iterable-iterator/map';
-
 import {useLocale} from '../i18n/datetime';
-import {settings} from './settings/hooks';
 
 import CustomWholeWindowDropZone from './input/CustomWholeWindowDropZone';
-import Header from './Header';
-import Content from './Content';
-import NavigationDrawer from './NavigationDrawer';
 import ErrorBoundary from './ErrorBoundary';
-
-import useLoggingIn from './users/useLoggingIn';
-import useUser from './users/useUser';
+import AppFrame from './AppFrame';
 
 let muitheme = createMuiTheme();
 muitheme = responsiveFontSizes(muitheme);
 
-const useStyles = makeStyles(() => ({
-	appFrame: {
-		position: 'relative',
-		display: 'flex',
-		width: '100%',
-		minHeight: '100vh',
-	},
-	progress: {
-		display: 'block',
-		position: 'fixed',
-		width: '100%',
-		zIndex: 9999,
-	},
-}));
-
-const useUISettings = () =>
-	useTracker(() => {
-		const handles = [
-			settings.subscribe('text-transform'),
-			settings.subscribe('navigation-drawer-is-open'),
-		];
-		return {
-			loading: !all(map((x) => x.ready(), handles)),
-			textTransform: settings.getWithBrowserCache('text-transform'),
-			navigationDrawerIsOpen: settings.getWithBrowserCache(
-				'navigation-drawer-is-open',
-			),
-		};
-	}, []);
-
 const App = () => {
-	const classes = useStyles();
-	const {loading, textTransform, navigationDrawerIsOpen} = useUISettings();
-	const loggingIn = useLoggingIn();
-	const currentUser = useUser();
 	const locale = useLocale();
 
 	return (
@@ -84,23 +36,7 @@ const App = () => {
 						<ErrorBoundary>
 							<div>
 								<CustomWholeWindowDropZone />
-								<div className={classes.appFrame} style={{textTransform}}>
-									{loading && <LinearProgress className={classes.progress} />}
-									<Header
-										navigationDrawerIsOpen={navigationDrawerIsOpen}
-										currentUser={currentUser}
-									/>
-									<NavigationDrawer
-										navigationDrawerIsOpen={navigationDrawerIsOpen}
-										currentUser={currentUser}
-									/>
-									<Content
-										navigationDrawerIsOpen={navigationDrawerIsOpen}
-										loggingIn={loggingIn}
-										currentUser={currentUser}
-										loading={loading}
-									/>
-								</div>
+								<AppFrame />
 							</div>
 						</ErrorBoundary>
 					</SnackbarProvider>
